# Conflicts:
#	dummy/gitrepo/components
#	pom.xml
#	src/main/java/com/ownspec/center/configuration/OsCenterConfiguration.java
#	src/main/resources/application.properties

+++ src/main/java/com/ownspec/center/model/Quantifiable.java
@@ -10,5 +10,5 @@ import javax.persistence.Embeddable;
 @Data
 @Embeddable
 public class Quantifiable {
-    private int hours;
+    private Double hours;
 }
